Rename "drawer" files to "footer" files
